Add EMCC flag for compiling with wasm simd support
Add the `-msimd128` and `-O2` flags to `EMCC_FLAGS` to compile with
auto-vectorization and WASM SIMD support.

https://webassembly.github.io/spec/core/exec/instructions.html#vector-instructions
https://v8.dev/features/simd

This breaks compatibility with Node versions < 16.4, however Node 14 LTS
rolls out of support April 2023, and supporting these older versions
isn't a priority.

I can confirm that SIMD instructions are used in the resulting wasm and
that there appears to be a speedup. I will follow this with an actual
benchmark.

Bump .nvmrc Node version to 18