Change `.getMime` to `.detect` (#38)
This was done for a couple of reasons. As features have been added to
WASMagic, the method name `.getMime` is no longer accurate. The user
can get any number of available features from the passed `Buffer`.
Second, consistency. Several other `libmagic` based libraries used the
`.detect` method (though the signatures differ).

While on the surface this looks like a BREAKING CHANGE, it is not. The
original `.getMime` method passes through to `.detect`. Though, it has
been marked as deprecated.