Fix StdioOverrideFunction Type export
This was a case of bad management of the types in play and their
exposure outside the package. The fix that works for both package
maintenance, and external users is to ensure that the type is available
and copied into the final npm package. Finally this would have been
caught with a test, which has now been added.

- Fix #40
- Add dedicated type file for StdioOverrideFunction
- Change type file paths to be relative
- Move all type files to the `src` folder to make things easier
- Add Makefile rules to copy over `.d.ts` files so they are in the
  package